moved run_nougat and make_todo_list to example.py
