+++ plotting/utils.py
@@ -9,50 +9,6 @@ import numpy as np
 import warnings
 
 
-def make_todo_list(quantities):
-    """
-    Make the todo_list for the Membrane object.
-
-    Parameters
-    ----------
-    quantities : str
-        A string specifying which quantities should be analyzed by nougat.py.
-
-    Raises
-    ------
-    ValueError
-        The user must specify a valid nougat quantity.
-
-    Returns
-    -------
-    todo_list  :  list
-        A list of nougat Field_sets to be computed.
-
-    """
-    if quantities is None:
-        todo_list = ["height", "curvature", "thickness", "order"]
-    else:
-        todo_list = []
-        for letter in quantities:
-            if letter == "h":
-                todo_list.append("height")
-            elif letter == "c":
-                todo_list.append("curvature")
-            elif letter == "t":
-                todo_list.append("thickness")
-            elif letter == "o":
-                todo_list.append("order")
-            elif letter == "n":
-                return NotImplemented
-                # todo_list.append("tilt")
-            elif letter == "d":
-                return NotImplemented
-                # todo_list.append("density")
-            else:
-                raise ValueError("Must specify a valid nougat quantity")
-    return todo_list
-
-
 def strip_blank_lines(file):
     """
     Remove any lines that are empty when reading a file.
